# api_docs.yaml
openapi: 3.0.0
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value 
info:
  title: Flashcard v2 Open API Documentation
  version: 1.0.0
  description: |
    <h2>ER Diagram</h2>
    [![](https://mermaid.ink/img/pako:eNqVU8tuhDAM_JXI591DHyfObX-gV6SVRQyNBAlyHK1WwL83EKpdSqjUcCDxTOwZyxmgcpqgAOI3gw1jV1oVV_DEXg3pMC9jhRriBbgYfQe8sLGNstjRLtij91fHezZ1aNoUndKvQtZeTeP5PI7xEIs5vqlClRC_B0ZG0gzkJNXsrOzpaP2VOJcmFc2lcr0YZy9PR8DzEfByBLz-cr8aHv6nK9t1Tb5is5TZFPGVY_q7x1tKGoIMnlG5IBuJhxOzuXIP6yjow3GHsmzFPFo7bEb0ljZwgo7ibaPjNC8CS5AviklgdqCpxtDKbGSK1NDPNd61EcdQ1Nh6OgEGcZ83W0EhHOiHtL6LlTV9A2pgAUw)](https://mermaid.live/edit#pako:eNqVU8tuhDAM_JXI591DHyfObX-gV6SVRQyNBAlyHK1WwL83EKpdSqjUcCDxTOwZyxmgcpqgAOI3gw1jV1oVV_DEXg3pMC9jhRriBbgYfQe8sLGNstjRLtij91fHezZ1aNoUndKvQtZeTeP5PI7xEIs5vqlClRC_B0ZG0gzkJNXsrOzpaP2VOJcmFc2lcr0YZy9PR8DzEfByBLz-cr8aHv6nK9t1Tb5is5TZFPGVY_q7x1tKGoIMnlG5IBuJhxOzuXIP6yjow3GHsmzFPFo7bEb0ljZwgo7ibaPjNC8CS5AviklgdqCpxtDKbGSK1NDPNd61EcdQ1Nh6OgEGcZ83W0EhHOiHtL6LlTV9A2pgAUw)
servers:
  - url: http://127.0.0.1:8080/
paths:
...
  /api/score:
     description: Scores endpoint
     get:
       description: Operation to get leaderboard details
       security:
         - bearerAuth: []
       responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                title: user object
                type: object
                properties:
                  leaderboard:
                    type: object
     post:
       description: Record a score
       security:
        - bearerAuth: []
       requestBody:
         content:
           application/json:
             schema:
               title: user object
               type: object
               properties:
                 category_id:
                   type: integer
                 score:
                   type: integer
       responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                title: user object
                type: object
                properties:
                  msg:
                    type: string
                    example: Scores stored successfully!
...